+++ Assets/Scripts/Debug/FPSManager/FPSManager.cs
@@ -265,7 +265,9 @@ public class FPSManager : Manager
 
         string dateTime = $"{DateTime.Now.ToLongDateString()} {DateTime.Now.ToShortTimeString()}";
         string operatingSystem = $"{SystemInfo.operatingSystem}";
-        string settings = $"{go.width}x{go.height}@{go.refreshRate}Hz ({go.fullScreenMode}) {o.screenPercentage}% SP ({mPix.ToString("F2")} MegaPixels)- {QualitySettings.names[QualitySettings.GetQualityLevel()]} Quality";
+        string sp = o.screenPercentage == 100 ? $"Native" : $"{o.screenPercentage}% SP ({o.upsamplingMethod.ToString()})";
+
+        string settings = $"{go.width}x{go.height}@{go.refreshRate}Hz ({go.fullScreenMode}) {sp} ({mPix.ToString("F2")} MegaPixels)- {QualitySettings.names[QualitySettings.GetQualityLevel()]} Quality";
         string bestFPS = $"{(1000 / min).ToString("F1")}fps ({min.ToString("F2")}ms)";
         string worstFPS = $"{(1000 / max).ToString("F1")}fps ({max.ToString("F2")}ms)";
         string averageFPS = $"{(1000 / med).ToString("F1")}fps ({med.ToString("F2")}ms)";